open one question at a time
