+++ src/routes/+page.svelte
@@ -4,7 +4,7 @@
 	import { userConnected, xmtpConnected, userState } from '$lib/stores/Network';
 	import Skeleton from '$lib/components/common/Skeleton.svelte';
 	import BenefitsCards from '$lib/components/landing/BenefitsCards.svelte';
-	import Faq from '$lib/components/landing/faq.svelte';
+	import Faq from '$lib/components/landing/FAQ.svelte';
 
 	const track = async () => {
 		if (typeof window !== 'undefined' && window) {
